requirements.txt
Django==4.1
django-cors-headers==4.3.0
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1
gunicorn==21.2.0
pandas==2.1.1
Pillow==10.0.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
XlsxWriter==3.1.9
pytest-django==4.5.2
pytest==7.4.3
coverage==7.3.2

